The source code for a little game I wrote back in 2005, in order to land my first programming job.
The engine is very, very bad, but I found the source code on an old hard-drive, and it seemed like a shame to risk losing it at some point (as has happened with some other codebases that have outlived their usefulness as my skills have improved over the years). Begrudgingly, I felt that I should keep it around, as it is a part of my history.